requirements.txt
fastapi
uvicorn
rich
pillow